product-service/requirements.txt
django==4.1.7
djangorestframework==3.14.0
grpc-django==1.0.0
drf-spectacular==0.25.1
djangorestframework-simplejwt==5.2.2
git+https://github.com/profmcdan/django-grpc-framework.git@main
